Move transformation and sharedcode pkgs to mapper dir
