+++ internal/pkg/mapper/sharedcode/common.go
